refactor: moves history tx types to dedicated file
