+++ src/models/transaction.ts
@@ -44,49 +44,6 @@ type optionsType = {
   hash?: string | null,
 };
 
-export type HistoryTransactionOutput = {
-  value: number,
-  token_data: number,
-  script: string,
-  decoded: {
-    type?: string,
-    address?: string,
-    timelock?: number | null,
-  },
-  token: string,
-  spent_by?: string | null,
-  selected_as_input?: boolean,
-};
-
-export type HistoryTransactionInput = {
-  value: number,
-  token_data: number,
-  script: string,
-  decoded: {
-    type?: string,
-    address?: string,
-    timelock?: number | null,
-  },
-  token: string,
-  tx_id: string,
-  index: number,
-};
-
-export type HistoryTransaction = {
-  tx_id: string,
-  version: number,
-  weight: number,
-  timestamp: number,
-  is_voided: boolean,
-  nonce: number,
-  inputs: HistoryTransactionInput[],
-  outputs: HistoryTransactionOutput[],
-  parents: string[],
-  token_name?: string,
-  token_symbol?: string,
-  tokens: string[],
-}
-
 /**
  * Representation of a transaction with helper methods.
  *
