+++ src/utils/helpers.ts
@@ -9,7 +9,8 @@ import { OP_PUSHDATA1 } from '../opcodes';
 import { DECIMAL_PLACES, DEFAULT_TX_VERSION, CREATE_TOKEN_TX_VERSION } from '../constants';
 import buffer from 'buffer';
 import Long from 'long';
-import Transaction, {HistoryTransaction, HistoryTransactionOutput} from '../models/transaction';
+import Transaction from '../models/transaction';
+import { HistoryTransaction, HistoryTransactionOutput } from '../models/types';
 import P2PKH from '../models/p2pkh';
 import P2SH from '../models/p2sh';
 import ScriptData from '../models/script_data';
